Phase 4 Audit Remediation: Decimal migration + full swarm fixes
Complete remediation of Phase 4 six-agent audit findings:

- H3: Full float→Decimal migration across 18 files (~200 touch points).
  All monetary fields, NPV/IRR calculations, commission rules, and
  currency conversions now use Decimal arithmetic. SQLite write paths
  convert Decimal→float for REAL columns; JSON boundary uses
  convert_to_json_safe for API output.

- H1, M1-M6, L1-L7: All lower-severity findings resolved — type safety
  fixes, DRY consolidation, import hygiene, UI layer compliance, error
  handling hardening, and naming standardization.

- New modules: card_models, file_archival, inbox_scan_service,
  sync_worker, detail_panel, file_card, inbox_card_view.

- BaseRepository: added _execute_with_fallback pattern and
  _queue_pending_sync for offline-first resilience.

Remaining: H2 (Pydantic payload models at service boundaries).

+++ app/models/fixed_cost.py
@@ -6,6 +6,7 @@ Fields derived from FixedCost() constructor calls in transactions.py.
 """
 
 from __future__ import annotations
+from decimal import Decimal
 from typing import Optional
 from pydantic import BaseModel, Field
 
@@ -21,10 +22,10 @@ class FixedCost(BaseModel):
     tipo_servicio: Optional[str] = None
     ticket: Optional[str] = None
     ubicacion: Optional[str] = None
-    cantidad: Optional[float] = Field(default=None, ge=0)
-    costo_unitario_original: Optional[float] = Field(default=None, ge=0)
+    cantidad: Optional[Decimal] = Field(default=None, ge=0)
+    costo_unitario_original: Optional[Decimal] = Field(default=None, ge=0)
     costo_unitario_currency: Currency = Currency.USD
-    costo_unitario_pen: Optional[float] = Field(default=None, ge=0)
+    costo_unitario_pen: Optional[Decimal] = Field(default=None, ge=0)
     periodo_inicio: int = Field(default=0, ge=0)
     duracion_meses: int = Field(default=1, ge=1)
